bok/requirements.txt
numpy
opencv_python
pymongo
python-dotenv